更新内容： 1. 重构了UserPermission控制器，以及业务逻辑。 2. siae-auth已经验证完毕，初步完工，后续测试后再进行修改。

+++ siae-web-starter/src/main/java/com/hngy/siae/web/utils/PageConvertUtil.java
@@ -130,11 +130,11 @@ public class PageConvertUtil {
      * @param <T>      数据类型
      * @return 空的 PageVO
      */
-    public static <T> PageVO<T> empty(int pageNum, int pageSize) {
+    public static <T> PageVO<T> empty(long pageNum, long pageSize) {
         PageVO<T> result = new PageVO<>();
         result.setTotal(0L);
-        result.setPageNum(pageNum);
-        result.setPageSize(pageSize);
+        result.setPageNum((int)pageNum);
+        result.setPageSize((int)pageSize);
         result.setRecords(List.of());
         return result;
     }
